AK: Make Bitmap use size_t for its size
Also rework its API's to return Optional<size_t> instead of int with -1
as the error value.

+++ Kernel/VM/Region.cpp
@@ -164,7 +164,7 @@ u32 Region::cow_pages() const
     if (!m_cow_map)
         return 0;
     u32 count = 0;
-    for (int i = 0; i < m_cow_map->size(); ++i)
+    for (size_t i = 0; i < m_cow_map->size(); ++i)
         count += m_cow_map->get(i);
     return count;
 }
